stop a compilation warning on ubuntu.
git-svn-id: http://svn.code.sf.net/p/gnomesword/code/trunk@3745 da23c209-522a-4b7c-92e9-24af42be4764

+++ src/main/xml.c
@@ -1184,7 +1184,7 @@ void xml_save_settings_doc(char *name)
 	if (((retval = stat(name, &buf)) < 0) ||
 	    ((retval == 0) && (buf.st_size == 0))) {
 		char *msg = g_strdup_printf(_("Save of settings failed! stat %d, size %d\n%s"),
-					    retval, buf.st_size,
+					    retval, (int)buf.st_size,
 					    _("Attempting to revert to previous save."));
 		unlink(name);
 		rename(backup_name, name);
